Successfully pushing correct history.jsonp to s3
