+++ src/utils/getPathFromGlob.js
@@ -14,7 +14,9 @@ function throwErrorsIfNeeded(globbedPaths, shortPath) {
 }
 
 module.exports = async function getPathFromGlob(file) {
+  console.log(`Finding path for ${file.fullPath}`);
   const globbedPaths = await globPromise(file.fullPath);
+
   throwErrorsIfNeeded(globbedPaths, file.shortPath);
   return globbedPaths[0];
 };
